+++ SOLVER/meshes_io.f90
@@ -111,12 +111,12 @@ subroutine dump_xdmf_grid()
   use nc_routines,      only: nc_dump_snap_points, nc_dump_snap_grid, nc_make_snapfile
   use data_numbering
 
-  integer              :: iel, ipol, jpol, ipol1, jpol1, i, j, ct, ipt, idest
-  real(4), allocatable :: points(:,:)
-  integer, allocatable :: grid(:,:), mapping(:)
-  logical, allocatable :: check(:), mask_tp_elem(:)
-  character(len=120)   :: fname
-  double precision     :: rmin, rmax, thetamin, thetamax
+  integer               :: iel, ipol, jpol, ipol1, jpol1, i, j, ct, ipt, idest
+  real(sp), allocatable :: points(:,:)
+  integer, allocatable  :: grid(:,:), mapping(:)
+  logical, allocatable  :: check(:), mask_tp_elem(:)
+  character(len=120)    :: fname
+  real(kind=dp)           :: rmin, rmax, thetamin, thetamax
   
   inquire(file="inparam_xdmf", EXIST=file_exists)
 
@@ -593,8 +593,8 @@ subroutine dump_wavefields_mesh_1d
   use data_pointwise
   use nc_routines, ONLY: nc_dump_mesh_sol, nc_dump_mesh_flu
   
-  double precision, dimension(:,:,:), allocatable :: ssol, zsol
-  double precision, dimension(:,:,:), allocatable :: sflu, zflu
+  real(kind=dp)    , dimension(:,:,:), allocatable :: ssol, zsol
+  real(kind=dp)    , dimension(:,:,:), allocatable :: sflu, zflu
   
   integer :: iel, ipol,jpol,i
   
@@ -743,7 +743,7 @@ subroutine fldout_cyl2(fname,nel,f,ibeg,iend,jbeg,jend,flag_norm,domain)
   integer, intent(in)             :: flag_norm,ibeg,iend,jbeg,jend,nel
   real(kind=realkind), intent(in) :: f(ibeg:iend,jbeg:jend,nel)
   integer                         :: lf,ielem, ipol,jpol,iel
-  double precision                :: r, theta, s, z
+  real(kind=dp)                    :: r, theta, s, z
   real(kind=realkind)             :: fnr,afnr
   
   lf=index(fname,' ')-1
